Devices path now based on a fixed folder via meta GUID
When Unity Device Simulator is installed or there are folder with .device in user's project, it was stealing attention and causing Notch Simulator to use those instead of our device database. Also we cannot use Unity Device Sim definitions since they came with outer paddings, it is better that we keep using our own folder. (While open our folder for Unity Device Sim to include) However we cannot hide the folder with "." or "~" because AssetDatabase need to index it.